Linked-Lists
The file stack.c was an assignment demonstrating the uses of pointers, structures, and unions. the stack type is the node of a linked list with a data value of either a string of max 10 characters or an integer. A union structure was used for the data of each node.